Add files via upload
Since 2008, guests and hosts have used Airbnb to expand on traveling possibilities and present more personalized ways of experiencing the world. This dataset contains information on 2019 listings in New York and its geographical information, prices, number of reviews, and moreWorked on:
Which hosts are the busiest and why?
What areas have more traffic than others and why is that the case?
Are there any relationships between prices, number of reviews, and the number of days that a given listing is booked?